add guess.rs
